upgrade accord dependency
accord depends on fobject.
fobject depended on an old version of graceful-fs that wasn't compatible with node 6, but this has been fixed now in both fobject and accord.

this is untested and submitted through the GH website, but should work fine?

accord ^0.22.5

This isn't released yet... but will be the next release that fixes this issue.

released as 0.23.0

add node 6 to travis testing

upgrade mocha dependency

also has an issue with glob depending on graceful-fs. sigh.